fix: change to es6 imports feat: add admin middleware
