+++ api/admin/adminMiddleware.js
@@ -1,0 +1,4 @@
+export default function (req, res, next) {
+  if (!req.user.isAdmin) return res.status(403).send('Access denied');
+  next();
+}
